asm: reimplement commit e260b036 to allow self-hosting with tcc 0.9.26
Commit e260b036 added support for 'setb' test variants using a macro which is
used with an empty argument for non 'set' tests. It's valid and also builds fine
with current tcc, but apparently the latest official tcc (0.9.26 - tested on
windows) doesn't like it.

This uses alternative form for the macro, which produces warnings but
builds. The resulting executable can then be used to build tcc again
cleanly and with all tests passing.