asm: reimplement commit e260b036 to allow self-hosting with tcc 0.9.26
Commit e260b036 added support for 'setb' test variants using a macro which is
used with an empty argument for non 'set' tests. It's valid and also builds fine
with current tcc, but apparently the latest official tcc (0.9.26 - tested on
windows) doesn't like it.

So duplicate and split it into two macros: one with and one without suffix, and
use the one with suffix only where a suffix is actually required.