win: build: try to make a reproducible self-hosting buld
Depends only on an external CC (no AR).

Currently if the external CC is gcc or cl (msvc) or tcc built with one of those
two, then the resulting tcc and libtcc1.a are identical and reproducible.

However, if the external CC is the official pre-built windows tcc-0.9.26 binary
package (32 or 64), then the resulting tcc.exe (and libtcc1.o and as a result
libtcc1.a) are not identical to those which gcc/cl produce. Trying again with
the resulting binaries will stay identical to the previous binaries (and
different than the ones which gcc/cl end up creating).

The differences between the two "classes" of binaries are small - they have
identical size, but on few (probably less than 5-10) occasions, the tcc-0.9.26
produced binary has a byte of 128 where the gcc/cl produced binary has 0.

This difference in the produced binaries doesn't affect all source files. E.g.
in libtcc1, only libtcc1.c produces an different object, but all the other
libtcc1 sources produce identical objects.

It's similar on other projects too. Some source files consistently produce
identical objects, while other sources don't - all with the same 0/128 diff.