win: reproducible-build: improvements
Currently it produces bit-identical win32 dist package of tcc when the
initial compiler is a recent tcc.exe, or gcc (tested msys2-32) or cl.exe
(tested msvc 2013).

However, when the initial compiler is from the official tcc 0.9.26 win32
package, the resulting build has few different bits, and no number of
more builds with the result of the previous build changes that.